+++ src/kernel/memory.hpp
@@ -27,11 +27,11 @@ public:
     void free(void* ptr);
 };
 
-void* operator new(unsigned size);
-void* operator new[](unsigned size);
+void* operator new(size_t size);
+void* operator new[](size_t size);
 
-void* operator new(unsigned size, void* ptr);
-void* operator new[](unsigned size, void* ptr);
+void* operator new(size_t size, void* ptr);
+void* operator new[](size_t size, void* ptr);
 
 void operator delete(void* ptr);
 void operator delete[](void* ptr);
