+++ src/libraries/LibC/types.hpp
@@ -11,6 +11,6 @@ typedef unsigned int uint32_t;
 typedef long long int int64_t;
 typedef unsigned long long int uint64_t;
 
-typedef uint32_t size_t;
+typedef long unsigned int size_t;
 
 #endif
